Port ACF 6.8 JS assets
