+++ assets/src/js/pro/_acf-blocks-v3.js
@@ -3,11 +3,11 @@
  * Imports all modules and initializes the block system
  */
 
-// Import utilities
-// import './blocks-v3/utils/post-locking';
+// Import utilities.
+import './blocks-v3/utils/post-locking';
 
-// Import JSX parser and expose on acf global
-// import './blocks-v3/components/jsx-parser';
+// Import JSX parser and expose on acf global.
+import './blocks-v3/components/jsx-parser';
 
-// Import block registration (initializes automatically via acf.addAction)
-// import './blocks-v3/register-block-type-v3';
+// Import block registration (initializes automatically via acf.addAction).
+import './blocks-v3/register-block-type-v3';
