+++ assets/src/js/pro/blocks-v3/components/block-edit.js
@@ -18,7 +18,13 @@ import {
 	useBlockProps,
 	useBlockEditContext,
 } from '@wordpress/block-editor';
-import { Button, Placeholder, Spinner, Modal } from '@wordpress/components';
+import {
+	Button,
+	Placeholder,
+	Spinner,
+	Modal,
+	PanelBody,
+} from '@wordpress/components';
 import { BlockPlaceholder } from './block-placeholder';
 import { BlockForm } from './block-form';
 import { BlockPreview } from './block-preview';
@@ -52,6 +58,32 @@ const InspectorBlockFormContainer = ( {
 	}, [] );
 
 	return <div ref={ inspectorBlockFormRef } />;
+};
+
+const isBlockEditorInspectorSidebarOpen = () => {
+	return (
+		wp.data
+			.select( 'core/interface' )
+			.getActiveComplementaryArea( 'core' ) === 'edit-post/block'
+	);
+};
+
+const useBlockEditorInspectorSidebarOpen = () => {
+	const [ isOpen, setIsOpen ] = useState( () =>
+		isBlockEditorInspectorSidebarOpen()
+	);
+
+	useEffect( () => {
+		const unsubscribe = wp.data.subscribe( () => {
+			setIsOpen( isBlockEditorInspectorSidebarOpen() );
+		} );
+
+		return () => {
+			unsubscribe();
+		};
+	}, [] );
+
+	return isOpen;
 };
 
 /**
@@ -88,7 +120,9 @@ export const BlockEdit = ( props ) => {
 
 	const [ showValidationErrors, setShowValidationErrors ] = useState( null );
 	const [ theSerializedAcfData, setTheSerializedAcfData ] = useState( null );
-	const [ blockFormHtml, setBlockFormHtml ] = useState( '' );
+	const [ blockFormHtml, setBlockFormHtml ] = useState(
+		() => preloadedData?.form ?? ''
+	);
 	const [ blockPreviewHtml, setBlockPreviewHtml ] = useState( () => {
 		if ( preloadedData?.html ) {
 			return acf.applyFilters(
@@ -103,10 +137,13 @@ export const BlockEdit = ( props ) => {
 		useState( false );
 	const [ hasFetchedOnce, setHasFetchedOnce ] = useState( false );
 	const [ ajaxRequest, setAjaxRequest ] = useState();
+	const [ isFetchingBlock, setIsFetchingBlock ] = useState( false );
 
 	// New state for inline editing features
 	const [ blockToolbarFields, setBlockToolbarFields ] = useState( [] );
-	const [ blockFieldInfo, setBlockFieldInfo ] = useState( null );
+	const [ blockFieldInfo, setBlockFieldInfo ] = useState(
+		() => preloadedData?.fields ?? null
+	);
 	const [ gutenbergIframeOrDocument, setGutenbergIframeOrDocument ] =
 		useState( () => {
 			const iframe = document.querySelector( '[name="editor-canvas"]' );
@@ -130,6 +167,12 @@ export const BlockEdit = ( props ) => {
 	] = useState( false );
 	const [ acfDynamicStylesElement, setAcfDynamicStylesElement ] =
 		useState( null );
+	const [
+		freezeInlineToolbarDuringReRender,
+		setFreezeInlineToolbarDuringReRender,
+	] = useState( '' );
+	const blockEditorInspectorSidebarOpen =
+		useBlockEditorInspectorSidebarOpen();
 
 	const acfFormRef = useRef( null );
 	const previewRef = useRef( null );
@@ -186,6 +229,8 @@ export const BlockEdit = ( props ) => {
 
 		if ( preloadedData ) {
 			handlePreloadedData( preloadedData );
+			unlockPostSavingByName( 'acf-fetching-block' );
+			setIsFetchingBlock( false );
 			return;
 		}
 
@@ -201,6 +246,7 @@ export const BlockEdit = ( props ) => {
 		const blockData = { ...theAttributes };
 
 		lockPostSavingByName( 'acf-fetching-block' );
+		setIsFetchingBlock( true );
 
 		// Fetch block data via AJAX
 		const request = $.ajax( {
@@ -218,6 +264,7 @@ export const BlockEdit = ( props ) => {
 		} )
 			.done( ( response ) => {
 				unlockPostSavingByName( 'acf-fetching-block' );
+				setIsFetchingBlock( false );
 
 				setBlockFormHtml( response.data.form );
 
@@ -264,6 +311,7 @@ export const BlockEdit = ( props ) => {
 			.fail( function () {
 				setHasFetchedOnce( true );
 				unlockPostSavingByName( 'acf-fetching-block' );
+				setIsFetchingBlock( false );
 			} );
 		setAjaxRequest( request );
 	}
@@ -309,6 +357,7 @@ export const BlockEdit = ( props ) => {
 
 		// Replace placeholder client ID with actual client ID
 		data.html = data.html.replaceAll( hash, clientId );
+		data.form = data.form.replaceAll( hash, clientId );
 
 		if ( data?.validation && data?.validation.errors ) {
 			data.validation.errors = data.validation.errors.map( ( error ) => {
@@ -347,8 +396,10 @@ export const BlockEdit = ( props ) => {
 	 */
 	function handlePreloadedData( data ) {
 		if ( data.form ) {
-			setBlockFormHtml( data.html );
-		} else if ( data.html ) {
+			setBlockFormHtml( data.form );
+		}
+
+		if ( data.html ) {
 			setBlockPreviewHtml(
 				acf.applyFilters( 'blocks/preview/render', data.html, true )
 			);
@@ -391,13 +442,6 @@ export const BlockEdit = ( props ) => {
 			window.removeEventListener( 'keydown', trackUserInteraction );
 		}
 
-		fetchBlockData( {
-			theAttributes: attributes,
-			theClientId: clientId,
-			theContext: context,
-			isSelected: isSelected,
-		} );
-
 		window.addEventListener( 'click', trackUserInteraction );
 		window.addEventListener( 'keydown', trackUserInteraction );
 
@@ -406,6 +450,16 @@ export const BlockEdit = ( props ) => {
 			window.removeEventListener( 'keydown', trackUserInteraction );
 		};
 	}, [] );
+
+	useEffect( () => {
+		if ( isSelected ) {
+			return;
+		}
+
+		setCurrentInlineEditingElementUid( null );
+		setCurrentInlineEditingElement( null );
+		setCurrentContentEditableElement( null );
+	}, [ isSelected ] );
 
 	// Update hasAcfError attribute based on validation errors
 	useEffect( () => {
@@ -447,6 +501,7 @@ export const BlockEdit = ( props ) => {
 	useEffect( () => {
 		clearTimeout( debounceRef.current );
 		lockPostSavingByName( 'acf-fetching-block' );
+		setIsFetchingBlock( true );
 
 		debounceRef.current = setTimeout( () => {
 			const parsedData = JSON.parse( theSerializedAcfData );
@@ -484,6 +539,7 @@ export const BlockEdit = ( props ) => {
 		return () => {
 			clearTimeout( debounceRef.current );
 			unlockPostSavingByName( 'acf-fetching-block' );
+			setIsFetchingBlock( false );
 		};
 	}, [ theSerializedAcfData, attributesWithoutError ] );
 
@@ -509,6 +565,31 @@ export const BlockEdit = ( props ) => {
 			}
 		}
 	}, [ blockPreviewHtml ] );
+
+	useEffect( () => {
+		if (
+			( currentContentEditableElement &&
+				! inlineEditingToolbarHasFocus ) ||
+			! currentInlineEditingElement
+		) {
+			return;
+		}
+
+		const toolbar = document.querySelector( '.acf-inline-editing-toolbar' );
+
+		if ( toolbar?.style?.cssText ) {
+			setFreezeInlineToolbarDuringReRender(
+				toolbar.style.cssText.replaceAll( ';', '!important;' )
+			);
+		}
+
+		setTimeout( () => {
+			setCurrentInlineEditingElement( currentInlineEditingElement );
+			setTimeout( () => {
+				setFreezeInlineToolbarDuringReRender( null );
+			}, 0 );
+		}, 0 );
+	}, [ currentInlineEditingElement ] );
 
 	return (
 		<BlockEditInner
@@ -547,6 +628,11 @@ export const BlockEdit = ( props ) => {
 			}
 			acfDynamicStylesElement={ acfDynamicStylesElement }
 			setAcfDynamicStylesElement={ setAcfDynamicStylesElement }
+			blockEditorInspectorSidebarOpen={ blockEditorInspectorSidebarOpen }
+			freezeInlineToolbarDuringReRender={
+				freezeInlineToolbarDuringReRender
+			}
+			isFetchingBlock={ isFetchingBlock }
 		/>
 	);
 };
@@ -590,14 +676,19 @@ function BlockEditInner( props ) {
 		setContentEditableChangeInProgress,
 		acfDynamicStylesElement,
 		setAcfDynamicStylesElement,
+		blockEditorInspectorSidebarOpen,
+		freezeInlineToolbarDuringReRender,
+		isFetchingBlock,
 	} = props;
 
 	const { clientId } = useBlockEditContext();
 	const inspectorControlsRef = useRef();
-	const [ isModalOpen, setIsModalOpen ] = useState( false );
 	const [ blockFormModalOpen, setBlockFormModalOpen ] = useState( false );
 	const modalFormContainerRef = useRef();
 	const [ currentFormContainer, setCurrentFormContainer ] = useState();
+	const [ canRenderForm, setCanRenderForm ] = useState( false );
+	const [ invisibleBlockFormContainer, setInvisibleBlockFormContainer ] =
+		useState();
 
 	// Render counter for debugging
 	const renderCount = useRef( 0 );
@@ -629,24 +720,56 @@ function BlockEditInner( props ) {
 		setAcfDynamicStylesElement( styleElement );
 	}, [ gutenbergIframeOrDocument ] );
 
-	// Set current form container when modal opens
 	useEffect( () => {
-		if ( isModalOpen && modalFormContainerRef?.current ) {
-			setCurrentFormContainer( modalFormContainerRef.current );
-		}
-	}, [ isModalOpen, modalFormContainerRef ] );
+		let invisibleContainer = document.getElementById(
+			'invisible-acf-form-element'
+		);
 
-	// Update form container when inspector panel is available
+		if ( ! invisibleContainer ) {
+			invisibleContainer = document.createElement( 'div' );
+			invisibleContainer.id = 'invisible-acf-form-element';
+			invisibleContainer.style.display = 'none';
+			document.body.appendChild( invisibleContainer );
+		}
+
+		setInvisibleBlockFormContainer( invisibleContainer );
+	}, [ blockEditorInspectorSidebarOpen ] );
+
+	useEffect( () => {
+		if ( blockFormModalOpen && modalFormContainerRef?.current ) {
+			setCurrentFormContainer( modalFormContainerRef.current );
+			return;
+		}
+
+		setCurrentFormContainer(
+			blockEditorInspectorSidebarOpen
+				? inspectorControlsRef.current
+				: invisibleBlockFormContainer
+		);
+	}, [ blockFormModalOpen, modalFormContainerRef ] );
+
+	useEffect( () => {
+		if ( blockEditorInspectorSidebarOpen ) {
+			if ( ! blockFormModalOpen ) {
+				setCurrentFormContainer( inspectorControlsRef.current );
+			}
+			return;
+		}
+
+		setCurrentFormContainer( invisibleBlockFormContainer );
+	}, [
+		blockEditorInspectorSidebarOpen,
+		invisibleBlockFormContainer,
+		blockFormModalOpen,
+	] );
+
 	useEffect( () => {
 		if ( isSelected && inspectorControlsRef?.current ) {
-			setCurrentFormContainer( inspectorControlsRef.current );
+			setCanRenderForm( true );
 		} else if ( isSelected && ! inspectorControlsRef?.current ) {
-			// Wait for inspector to be available
 			setTimeout( () => {
-				setCurrentFormContainer( inspectorControlsRef.current );
+				setCanRenderForm( true );
 			}, 1 );
-		} else if ( ! isSelected ) {
-			setCurrentFormContainer( null );
 		}
 	}, [ isSelected, inspectorControlsRef, inspectorControlsRef.current ] );
 
@@ -657,7 +780,7 @@ function BlockEditInner( props ) {
 			showValidationErrors &&
 			blockType?.hide_fields_in_sidebar
 		) {
-			setIsModalOpen( true );
+			setBlockFormModalOpen( true );
 		}
 	}, [ isSelected, showValidationErrors, validationErrors, blockType ] );
 
@@ -839,12 +962,13 @@ function BlockEditInner( props ) {
 		}
 	};
 
-	// Determine portal target
-	let portalTarget = null;
+	let portalTarget =
+		blockEditorInspectorSidebarOpen && inspectorControlsRef?.current
+			? inspectorControlsRef.current
+			: invisibleBlockFormContainer;
+
 	if ( currentFormContainer ) {
 		portalTarget = currentFormContainer;
-	} else if ( inspectorControlsRef?.current ) {
-		portalTarget = inspectorControlsRef.current;
 	}
 
 	// Determine inline editing toolbar anchor (matches 6.7.0.2 logic)
@@ -880,12 +1004,22 @@ function BlockEditInner( props ) {
 				gutenbergIframeOrDocument={ gutenbergIframeOrDocument }
 				setBlockFormModalOpen={ setBlockFormModalOpen }
 				blockFormModalOpen={ blockFormModalOpen }
+				invisibleBlockFormContainer={ invisibleBlockFormContainer }
 				currentInlineEditingElement={ currentInlineEditingElement }
 				setCurrentInlineEditingElement={
 					setCurrentInlineEditingElement
 				}
 				currentInlineEditingElementUid={
 					currentInlineEditingElementUid
+				}
+				hideExpandedEditorBtnInToolbar={
+					blockType?.expanded_editor_buttons === false ||
+					( Array.isArray( blockType?.expanded_editor_buttons ) &&
+						! blockType?.expanded_editor_buttons.includes(
+							'toolbar'
+						) ) ||
+					! blockFieldInfo ||
+					blockFieldInfo?.length === 0
 				}
 				onNewInlineEditingElementSelected={
 					handleNewInlineEditingElementSelected
@@ -894,17 +1028,26 @@ function BlockEditInner( props ) {
 
 			{ /* Inspector panel container */ }
 			<InspectorControls>
-				<div style={ { padding: '16px' } }>
-					<Button
-						className="acf-blocks-open-expanded-editor-btn"
-						variant="secondary"
-						onClick={ () => {
-							setIsModalOpen( true );
-						} }
-						text={ acf.__( 'Open Expanded Editor' ) }
-						icon="edit"
-					/>
-				</div>
+				{ blockFieldInfo?.length > 0 &&
+					( blockType?.expanded_editor_buttons === true ||
+						( Array.isArray( blockType?.expanded_editor_buttons ) &&
+							blockType?.expanded_editor_buttons.includes(
+								'sidebar'
+							) ) ) && (
+						<PanelBody>
+							<Button
+								className="acf-blocks-open-expanded-editor-btn"
+								variant="secondary"
+								onClick={ () => {
+									setBlockFormModalOpen( true );
+								} }
+								icon="edit"
+							>
+								{ blockType?.expanded_editor_button_text ||
+									acf.__( 'Open Expanded Editor' ) }
+							</Button>
+						</PanelBody>
+					) }
 				<InspectorBlockFormContainer
 					inspectorBlockFormRef={ inspectorControlsRef }
 					setCurrentBlockFormContainer={ setCurrentFormContainer }
@@ -913,88 +1056,81 @@ function BlockEditInner( props ) {
 
 			{ /* Render form via portal when container is available */ }
 			{ portalTarget &&
+				canRenderForm &&
 				createPortal(
 					<>
 						<BlockForm
 							$={ $ }
 							clientId={ clientId }
 							blockFormHtml={ blockFormHtml }
-							onMount={ () => {
-								if ( ! hasFetchedOnce ) {
-									blockFetcher( {
-										theAttributes: attributes,
-										theClientId: clientId,
-										theContext: context,
-										isSelected: isSelected,
-									} );
-								}
-							} }
 							onChange={ function ( $form ) {
 								const serializedData = acf.serialize(
 									$form,
 									`acf-block_${ clientId }`
 								);
 								if ( serializedData ) {
-									// Normalize flexible content data for validation
-									const normalizedData =
-										acf.normalizeFlexibleContentData(
-											serializedData
-										);
 									setTheSerializedAcfData(
-										JSON.stringify( normalizedData )
+										JSON.stringify( serializedData )
 									);
 								}
 							} }
 							validationErrors={ validationErrors }
 							showValidationErrors={ showValidationErrors }
 							acfFormRef={ acfFormRef }
-							theSerializedAcfData={ theSerializedAcfData }
 							userHasInteractedWithForm={
 								userHasInteractedWithForm
 							}
-							setCurrentBlockFormContainer={
-								setCurrentFormContainer
-							}
 							attributes={ attributes }
 							hideFieldsInSidebar={
-								blockType?.hide_fields_in_sidebar &&
-								( ! currentFormContainer ||
+								( blockType?.auto_inline_editing &&
+									blockType?.hide_fields_in_sidebar ===
+										undefined &&
+									inspectorControlsRef.current ===
+										currentFormContainer ) ||
+								( blockType?.hide_fields_in_sidebar &&
 									inspectorControlsRef.current ===
 										currentFormContainer )
 							}
 						/>
+						{ freezeInlineToolbarDuringReRender && (
+							<style>
+								{ `.acf-inline-editing-toolbar{${ freezeInlineToolbarDuringReRender }}` }
+							</style>
+						) }
 					</>,
-					currentFormContainer || inspectorControlsRef.current
+					portalTarget
 				) }
 			<>
 				{ /* Modal for editing block fields */ }
-				{ isModalOpen && (
+				{ blockFormModalOpen && (
 					<Modal
 						className="acf-block-form-modal"
+						overlayClassName={ acf.applyFilters(
+							'blocks/expanded_editor_overlay_class',
+							'acf-expanded-editor-panel-overlay'
+						) }
 						isFullScreen={ true }
 						title={ blockType.title }
 						onRequestClose={ () => {
-							// Check if block is fetching
-							const isFetching =
-								document.body.classList.contains(
-									'acf-fetching-block'
-								);
-							if ( ! isFetching ) {
-								setCurrentFormContainer( null );
-								setIsModalOpen( false );
+							if ( ! isFetchingBlock || validationErrors ) {
+								setBlockFormModalOpen( false );
 							}
 						} }
+						shouldCloseOnEsc={
+							! isFetchingBlock || validationErrors
+						}
 						isDismissible={ false }
 						headerActions={ [
 							<Button
 								key="done"
 								variant="primary"
-								disabled={ document.body.classList.contains(
-									'acf-fetching-block'
-								) }
+								disabled={
+									isFetchingBlock && ! validationErrors
+								}
+								isBusy={ isFetchingBlock }
 								onClick={ () => {
 									setCurrentFormContainer( null );
-									setIsModalOpen( false );
+									setBlockFormModalOpen( false );
 								} }
 							>
 								{ acf.__( 'Done' ) }
@@ -1039,7 +1175,7 @@ function BlockEditInner( props ) {
 								! document.querySelector(
 									'.acf-inline-fields-popover-inner'
 								) &&
-								! inlineEditingToolbarHasFocus &&
+								! currentContentEditableElement &&
 								( currentInlineEditingElement &&
 									currentInlineEditingElement.focus(),
 								setCurrentInlineEditingElementUid( null ),
@@ -1107,10 +1243,12 @@ function BlockEditInner( props ) {
 				acfDynamicStylesElement &&
 				createPortal(
 					<style>
-						{ `[data-acf-inline-fields-uid="${ currentInlineEditingElementUid }"]{
-							outline: 2px solid var(--wp-admin-theme-color);
-							outline-offset: 2px;
-						}` }
+						{ `
+				[data-acf-inline-fields-uid="${ currentInlineEditingElementUid }"]{
+					outline: 2px solid var( --wp-admin-theme-color );
+					outline-offset: 2px;
+				}
+			` }
 					</style>,
 					acfDynamicStylesElement
 				) }
@@ -1127,33 +1265,14 @@ function BlockEditInner( props ) {
 								blockLabel={
 									blockType?.title || acf.__( 'ACF Block' )
 								}
-								setBlockFormModalOpen={ setIsModalOpen }
+								setBlockFormModalOpen={ setBlockFormModalOpen }
 								error={ error }
 							/>
 						) }
-						onError={ ( error, errorInfo ) => {
-							acf.debug(
-								'Block preview error caught:',
-								error,
-								errorInfo
-							);
-						} }
-						resetKeys={ [ blockPreviewHtml ] }
-						onReset={ ( { reason, next, prev } ) => {
-							acf.debug( 'Error boundary reset:', reason );
-							if ( reason === 'keys' ) {
-								acf.debug(
-									'Preview HTML changed from',
-									prev,
-									'to',
-									next
-								);
-							}
-						} }
 					>
 						{ blockPreviewHtml === 'acf-block-preview-no-html' ? (
 							<BlockPlaceholder
-								setBlockFormModalOpen={ setIsModalOpen }
+								setBlockFormModalOpen={ setBlockFormModalOpen }
 								blockLabel={ blockType.title }
 							/>
 						) : null }
@@ -1169,7 +1288,14 @@ function BlockEditInner( props ) {
 						{ blockPreviewHtml !== 'acf-block-preview-loading' &&
 							blockPreviewHtml !== 'acf-block-preview-no-html' &&
 							blockPreviewHtml &&
-							acf.parseJSX( blockPreviewHtml, $ ) }
+							acf.parseJSX(
+								blockPreviewHtml,
+								handleNewInlineEditingElementSelected,
+								updateFieldValueFromContentEditable,
+								handleNewContentEditableElementSelected,
+								blockFieldInfo,
+								$
+							) }
 					</ErrorBoundary>
 				</BlockPreview>
 			</>
