+++ assets/src/js/pro/blocks-v3/components/block-form.js
@@ -14,7 +14,6 @@ import { lockPostSaving, unlockPostSaving } from '../utils/post-locking';
  * @param {jQuery} props.$ - jQuery instance
  * @param {string} props.clientId - Block client ID
  * @param {string} props.blockFormHtml - HTML markup for the ACF form
- * @param {Function} props.onMount - Callback when form is mounted
  * @param {Function} props.onChange - Callback when form data changes
  * @param {Array} props.validationErrors - Array of validation error objects
  * @param {boolean} props.showValidationErrors - Whether to display validation errors
@@ -27,7 +26,6 @@ export const BlockForm = ( {
 	$,
 	clientId,
 	blockFormHtml,
-	onMount,
 	onChange,
 	validationErrors,
 	showValidationErrors,
@@ -40,35 +38,21 @@ export const BlockForm = ( {
 	const [ pendingChange, setPendingChange ] = useState( false );
 	const debounceTimer = useRef( null );
 	const [ userInteracted, setUserInteracted ] = useState( false );
-	const [ initialValuesCaptured, setInitialValuesCaptured ] =
-		useState( false );
-
-	// Call onMount when component first mounts
-	useEffect( () => {
-		onMount();
-	}, [] );
+	const [ hasRemounted, setHasRemounted ] = useState( false );
 
 	// Trigger onChange when there's a pending change
 	useEffect( () => {
-		if ( pendingChange ) {
-			// For the first change, capture default values even without interaction
-			if (
-				! initialValuesCaptured ||
-				userHasInteractedWithForm ||
-				userInteracted
-			) {
-				onChange( pendingChange );
-				setPendingChange( false );
-				if ( ! initialValuesCaptured ) {
-					setInitialValuesCaptured( true );
-				}
-			}
+		if (
+			pendingChange &&
+			( userHasInteractedWithForm || userInteracted )
+		) {
+			onChange( pendingChange );
+			setPendingChange( false );
 		}
 	}, [
 		pendingChange,
 		userHasInteractedWithForm,
 		userInteracted,
-		initialValuesCaptured,
 		setPendingChange,
 		onChange,
 	] );
@@ -183,10 +167,10 @@ export const BlockForm = ( {
 		const $form = $( formElement );
 		let isActive = true;
 
-		acf.doAction( 'remount', $form );
-		if ( ! initialValuesCaptured ) {
-			onChange( $form );
-			setInitialValuesCaptured( true );
+		if ( ! hasRemounted ) {
+			acf.debug( 'Remounting ACF Form' );
+			acf.doAction( 'remount', $form );
+			setHasRemounted( true );
 		}
 
 		const handleChange = () => {
@@ -202,6 +186,8 @@ export const BlockForm = ( {
 			inputs.forEach( ( input ) => {
 				input.removeEventListener( 'input', handleChange );
 				input.addEventListener( 'input', handleChange );
+				input.removeEventListener( 'change', handleChange );
+				input.addEventListener( 'change', handleChange );
 			} );
 
 			selects.forEach( ( select ) => {
@@ -252,6 +238,7 @@ export const BlockForm = ( {
 			.querySelectorAll( 'input, textarea' )
 			.forEach( ( input ) => {
 				input.addEventListener( 'input', handleChange );
+				input.addEventListener( 'change', handleChange );
 			} );
 
 		formElement.querySelectorAll( 'select' ).forEach( ( select ) => {
@@ -270,6 +257,7 @@ export const BlockForm = ( {
 					.querySelectorAll( 'input, textarea' )
 					.forEach( ( input ) => {
 						input.removeEventListener( 'input', handleChange );
+						input.removeEventListener( 'change', handleChange );
 					} );
 
 				formElement
