+++ assets/src/js/pro/blocks-v3/components/jsx-parser.js
@@ -3,6 +3,9 @@
  * Converts HTML strings to React/JSX elements for rendering in the block editor
  */
 
+/* global acf, React, Text */
+
+// eslint-disable-next-line import/no-unresolved -- WordPress provides jquery as an external script dependency.
 import jQuery from 'jquery';
 import { createElement, createRef } from '@wordpress/element';
 
@@ -12,6 +15,9 @@ const useInnerBlocksProps =
 
 /**
  * Gets the JSX-compatible name for an HTML attribute
+ *
+ * @param {string} attrName Attribute name.
+ * @return {string} JSX-compatible attribute name.
  */
 function getJSXNameReplacement( attrName ) {
 	return acf.isget( acf, 'jsxNameReplacements', attrName ) || attrName;
@@ -19,6 +25,9 @@ function getJSXNameReplacement( attrName ) {
 
 /**
  * ACF InnerBlocks wrapper component
+ *
+ * @param {Object} props Component props.
+ * @return {JSX.Element} InnerBlocks wrapper element.
  */
 function ACFInnerBlocksComponent( props ) {
 	const { className = 'acf-innerblocks-container' } = props;
@@ -54,10 +63,15 @@ class ScriptComponent extends React.Component {
 
 /**
  * Helper function to parse style attribute string into object
+ *
+ * @param {string} styleString Style attribute value.
+ * @return {Object} Parsed style object.
  */
 function parseStyleAttribute( styleString ) {
 	const styleObj = {};
-	if ( ! styleString ) return styleObj;
+	if ( ! styleString ) {
+		return styleObj;
+	}
 
 	styleString.split( ';' ).forEach( ( rule ) => {
 		const colonIndex = rule.indexOf( ':' );
@@ -79,6 +93,9 @@ function parseStyleAttribute( styleString ) {
 /**
  * Parses and transforms a DOM attribute to React props format
  * Based on 6.7.0.2's implementation
+ *
+ * @param {Attr} attribute DOM attribute.
+ * @return {Object} React prop name and value.
  */
 function parseAttribute( attribute ) {
 	let attrName = attribute.name;
@@ -90,7 +107,9 @@ function parseAttribute( attribute ) {
 		false,
 		attribute
 	);
-	if ( customParsed ) return customParsed;
+	if ( customParsed ) {
+		return customParsed;
+	}
 
 	switch ( attrName ) {
 		case 'class':
@@ -113,7 +132,9 @@ function parseAttribute( attribute ) {
 			break;
 		default:
 			// Skip data- attributes processing, keep them as-is
-			if ( attrName.indexOf( 'data-' ) === 0 ) break;
+			if ( attrName.indexOf( 'data-' ) === 0 ) {
+				break;
+			}
 
 			attrName = getJSXNameReplacement( attrName );
 			const firstChar = attrValue.charAt( 0 );
@@ -131,8 +152,44 @@ function parseAttribute( attribute ) {
 /**
  * Main parseNodeToJSX function - Based on 6.7.0.2's function `O`
  * Recursively converts DOM nodes to React/JSX elements
+ *
+ * @param {Node}   node  DOM node.
+ * @param {number} depth Recursion depth.
+ * @return {JSX.Element|null} Parsed JSX element.
  */
-function parseNodeToJSX( node, depth = 0 ) {
+function parseNodeToJSX( node, depth = 0, callbacks = {} ) {
+	const {
+		onNewInlineEditingElementSelected,
+		onNewContentEditableElementSelected,
+		blockFieldInfo,
+	} = callbacks;
+
+	const selectInlineEditingElement = ( uid, element = null ) => {
+		if ( onNewInlineEditingElementSelected ) {
+			onNewInlineEditingElementSelected( uid );
+			return;
+		}
+
+		if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
+			acf.blockEdit.setCurrentInlineEditingElementUid( uid );
+		}
+
+		if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
+			acf.blockEdit.setCurrentInlineEditingElement( element );
+		}
+	};
+
+	const selectContentEditableElement = ( fieldSlug, element = null ) => {
+		if ( onNewContentEditableElementSelected ) {
+			onNewContentEditableElementSelected( fieldSlug );
+			return;
+		}
+
+		if ( acf.blockEdit?.setCurrentContentEditableElement ) {
+			acf.blockEdit.setCurrentContentEditableElement( element );
+		}
+	};
+
 	// Determine component type
 	const nodeName = node.nodeName.toLowerCase();
 	let componentType;
@@ -149,7 +206,9 @@ function parseNodeToJSX( node, depth = 0 ) {
 			componentType = getJSXNameReplacement( nodeName );
 	}
 
-	if ( ! componentType ) return null;
+	if ( ! componentType ) {
+		return null;
+	}
 
 	const props = {};
 
@@ -179,15 +238,7 @@ function parseNodeToJSX( node, depth = 0 ) {
 			const uid = node.attributes.getNamedItem(
 				'data-acf-inline-fields-uid'
 			).value;
-			if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-				acf.blockEdit.setCurrentInlineEditingElementUid( uid );
-			}
-			// Also set the element reference for the toolbar to use
-			if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
-				acf.blockEdit.setCurrentInlineEditingElement(
-					event.currentTarget
-				);
-			}
+			selectInlineEditingElement( uid, event.currentTarget );
 		};
 
 		props.onMouseDown = ( event ) => event.stopPropagation();
@@ -205,14 +256,7 @@ function parseNodeToJSX( node, depth = 0 ) {
 				const uid = node.attributes.getNamedItem(
 					'data-acf-inline-fields-uid'
 				).value;
-				if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-					acf.blockEdit.setCurrentInlineEditingElementUid( uid );
-				}
-				if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
-					acf.blockEdit.setCurrentInlineEditingElement(
-						event.currentTarget
-					);
-				}
+				selectInlineEditingElement( uid, event.currentTarget );
 			}
 		};
 
@@ -228,9 +272,7 @@ function parseNodeToJSX( node, depth = 0 ) {
 					const uid = node.attributes.getNamedItem(
 						'data-acf-inline-fields-uid'
 					).value;
-					if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-						acf.blockEdit.setCurrentInlineEditingElementUid( uid );
-					}
+					selectInlineEditingElement( uid, event.currentTarget );
 				}
 			}
 			if ( event.key === 'Enter' ) {
@@ -243,14 +285,7 @@ function parseNodeToJSX( node, depth = 0 ) {
 				const uid = node.attributes.getNamedItem(
 					'data-acf-inline-fields-uid'
 				).value;
-				if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-					acf.blockEdit.setCurrentInlineEditingElementUid( uid );
-				}
-				if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
-					acf.blockEdit.setCurrentInlineEditingElement(
-						event.currentTarget
-					);
-				}
+				selectInlineEditingElement( uid, event.currentTarget );
 			}
 		};
 	}
@@ -261,10 +296,10 @@ function parseNodeToJSX( node, depth = 0 ) {
 			'data-acf-inline-contenteditable-field-slug'
 		).value;
 
-		// Get block field info from state if available
-		const blockFieldInfo = acf.blockEdit?.getBlockFieldInfo?.();
-		const editableFields = blockFieldInfo
-			? blockFieldInfo.filter(
+		const fieldInfo =
+			blockFieldInfo || acf.blockEdit?.getBlockFieldInfo?.();
+		const editableFields = fieldInfo
+			? fieldInfo.filter(
 					( field ) =>
 						field.name === fieldSlug &&
 						( field.type === 'text' || field.type === 'textarea' )
@@ -284,19 +319,14 @@ function parseNodeToJSX( node, depth = 0 ) {
 					acf.debug( `Navigation prevented for ${ link.href }` );
 				}
 				event.stopPropagation();
+				selectContentEditableElement( fieldSlug, event.currentTarget );
 				if ( node.hasAttribute( 'data-acf-inline-fields' ) ) {
 					const uid = node.attributes.getNamedItem(
 						'data-acf-inline-fields-uid'
 					).value;
-					if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-						acf.blockEdit.setCurrentInlineEditingElementUid( uid );
-					}
-					if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
-						acf.blockEdit.setCurrentInlineEditingElement( event.currentTarget );
-					}
-				}
-				if ( acf.blockEdit?.setCurrentContentEditableElement ) {
-					acf.blockEdit.setCurrentContentEditableElement( event.currentTarget );
+					selectInlineEditingElement( uid, event.currentTarget );
+				} else {
+					selectInlineEditingElement( null );
 				}
 			};
 
@@ -320,15 +350,8 @@ function parseNodeToJSX( node, depth = 0 ) {
 	) {
 		props.onClick = ( event ) => {
 			if ( event.target === event.currentTarget ) {
-				if ( acf.blockEdit?.setCurrentInlineEditingElementUid ) {
-					acf.blockEdit.setCurrentInlineEditingElementUid( null );
-				}
-				if ( acf.blockEdit?.setCurrentInlineEditingElement ) {
-					acf.blockEdit.setCurrentInlineEditingElement( null );
-				}
-				if ( acf.blockEdit?.setCurrentContentEditableElement ) {
-					acf.blockEdit.setCurrentContentEditableElement( null );
-				}
+				selectInlineEditingElement( null );
+				selectContentEditableElement( null );
 			}
 		};
 	}
@@ -348,7 +371,9 @@ function parseNodeToJSX( node, depth = 0 ) {
 				elementArgs.push( textContent );
 			}
 		} else {
-			elementArgs.push( parseNodeToJSX( childNode, depth + 1 ) );
+			elementArgs.push(
+				parseNodeToJSX( childNode, depth + 1, callbacks )
+			);
 		}
 	} );
 
@@ -360,8 +385,39 @@ function parseNodeToJSX( node, depth = 0 ) {
 /**
  * Main parseJSX function exposed on the acf global object
  * Matches 6.7.0.2's implementation exactly
+ *
+ * @param {string}   htmlString                         HTML string.
+ * @param {Function} onNewInlineEditingElementSelected  Inline element selection callback or jQuery parser.
+ * @param {Function} onContentEditableChange            ContentEditable change callback.
+ * @param {Function} onNewContentEditableElementSelected ContentEditable selection callback.
+ * @param {Array}    blockFieldInfo                     Field metadata.
+ * @param {Function} $                                  jQuery-compatible parser.
+ * @return {JSX.Element|Array} Parsed JSX children.
  */
-export function parseJSX( htmlString, $ = jQuery ) {
+export function parseJSX(
+	htmlString,
+	onNewInlineEditingElementSelected = jQuery,
+	onContentEditableChange = null,
+	onNewContentEditableElementSelected = null,
+	blockFieldInfo = null,
+	$ = null
+) {
+	const isJQueryParser =
+		typeof onNewInlineEditingElementSelected === 'function' &&
+		onNewInlineEditingElementSelected.fn &&
+		onNewInlineEditingElementSelected.fn.jquery;
+	const parser = isJQueryParser
+		? onNewInlineEditingElementSelected
+		: $ || jQuery;
+	const callbacks = isJQueryParser
+		? {}
+		: {
+				onNewInlineEditingElementSelected,
+				onContentEditableChange,
+				onNewContentEditableElementSelected,
+				blockFieldInfo,
+		  };
+
 	// Wrap in div to ensure valid HTML structure
 	htmlString = '<div>' + htmlString + '</div>';
 
@@ -372,9 +428,26 @@ export function parseJSX( htmlString, $ = jQuery ) {
 	);
 
 	// Parse with jQuery, convert to React, and extract children from wrapper div
-	const parsedElement = parseNodeToJSX( $( htmlString )[ 0 ], 0 );
+	const parsedElement = parseNodeToJSX(
+		parser( htmlString )[ 0 ],
+		0,
+		callbacks
+	);
 	return parsedElement.props.children;
 }
 
-// Expose parseJSX function on acf global object for backward compatibility
-acf.parseJSX = parseJSX;
+// Preserve the legacy parser for v1/v2 blocks, which pass the block version
+// as the second argument.
+const legacyParseJSX = acf.parseJSX;
+
+acf.parseJSXV3 = parseJSX;
+acf.parseJSX = ( htmlString, parserOrBlockVersion = jQuery, ...args ) => {
+	if (
+		typeof parserOrBlockVersion === 'number' &&
+		typeof legacyParseJSX === 'function'
+	) {
+		return legacyParseJSX( htmlString, parserOrBlockVersion );
+	}
+
+	return parseJSX( htmlString, parserOrBlockVersion, ...args );
+};
